Move some layout out of layout.ts into Score/part
