+++ src/layout/elements/Score.ts
@@ -1,27 +1,17 @@
 import types from "..";
 import * as notation from "../../notation";
-import { AbstractGroup } from "../layouts/AbstractGroup";
+import { SimpleGroup } from "../layouts/SimpleGroup";
 import { Box } from "../utils";
 
-export const PAGE_MARGIN = 0.5;
-
-export class Score extends AbstractGroup<types.Part, "Score", null> implements types.Score {
+export class Score extends SimpleGroup<types.Part, "Score", null> implements types.Score {
   readonly type = "Score";
-  readonly elements: types.Part[] = [];
 
   constructor(readonly score: notation.Score) {
-    super(Box.empty());
-  }
-
-  addElement(element: types.Part): void {
-    this.elements.push(element);
+    super();
   }
 
   layout() {
-    for (const element of this.elements) {
-      element.layout?.();
-    }
-
-    this.box = Box.encompass(...this.elements.map((e) => e.box)).update({ x: 0, y: 0 });
+    super.layout();
+    this.box = Box.encompass(...this.elements.map((e) => e.box));
   }
 }
