+++ src/layout/layouts/FlexGroup.ts
@@ -30,7 +30,7 @@ export type FlexGroupConfig = {
 export abstract class FlexGroup<
   T extends types.LayoutElement,
   Type extends string = "Group",
-  Parent extends types.LayoutElement = types.LayoutElement
+  Parent extends types.LayoutElement | null = types.LayoutElement
 > {
   abstract readonly type: Type;
   readonly elements: T[] = [];
